Refactor consumer into services.
