+++ .gitignore
@@ -4,3 +4,5 @@
 .vscode
 evedata-server
 debug
+profile.out
+coverage.txt
