--- apps/monolith/.gitignore
@@ -1,2 +1,0 @@
-debug
-monolith
