+++ build.sh
@@ -1,11 +1,7 @@
 #!/bin/bash
 set -e
 
-GOMAXPROCS=1 go test -timeout 90s ./...
-GOMAXPROCS=4 go test -timeout 90s -race ./...
-
-# no tests, but a build is something
-for dir in apps/*/; do
+for dir in cmd/*/; do
     dir=${dir%/}
     if grep -q '^package main$' $dir/*.go 2>/dev/null; then
         echo "building $dir"
