+++ cmd/monolith/.gitignore
@@ -1,0 +1,2 @@
+debug
+monolith
