--- fpGrowth/fpGrowth_test.go
@@ -1,506 +1,0 @@
-package fpGrowth
-
-import "testing"
-
-func TestFPGrowthMulti(t *testing.T) {
-	x := NewFPTree(testData, 2)
-	x.Growth()
-}
-
-func TestFPGrowthSingle(t *testing.T) {
-	x := NewFPTree(testData, 5)
-	x.Growth()
-}
-
-var testData ItemSet = ItemSet{
-	1289334600: {2076108335, 1524822631, 1180475789, 1742120029, 1408018716, 1260231228, 408502361},
-	1291414740: {1376878485, 393585298, 586095311},
-	1308810000: {90895213, 90890609, 90831103, 90894280, 877310002, 90911153},
-	1310565480: {1180475789, 862891259, 849575930},
-	1316984760: {1798144414, 442364277, 90911153},
-	1337049000: {91988770, 90911153, 91685297, 408502361},
-	1337629740: {91685297, 90911153, 408502361, 91665661, 849575930},
-	1288740180: {1150633923, 1528806877, 1849507302},
-	1336681620: {695023501, 92000794, 92001064},
-	1349352300: {92132328, 92000794, 92368750, 92255489, 1136921030, 92459112},
-	1288837560: {142762673, 689053914, 201150810},
-	1288741260: {1053105037, 733670887, 539096076, 1333748060, 1982122495, 1146422975},
-	1292453880: {1087211105, 1576576772, 90061639},
-	1307878260: {90848368, 90813929, 1047101522, 90848599, 90848365, 90816602},
-	1313156040: {830235770, 90677279, 264407574},
-	1323903780: {91412151, 91412200, 91413008, 91413055},
-	1325690340: {90686534, 91240453, 91243871},
-	1327612680: {1360499003, 468575525, 91414542},
-	1329723900: {90411679, 689952182, 90981945, 1473103725, 681821034, 90621446, 90200742},
-	1330450680: {492419687, 91245379, 91651693},
-	1333635300: {91806764, 91214844, 91532973},
-	1336036380: {869228841, 1485875899, 90359584, 595793285},
-	1336127040: {90079784, 90413794, 90521899, 91578991, 91725964, 307980762, 91177690, 90887840, 90917237, 91697688, 1634464743},
-	1336350480: {90513070, 91280528, 91358586, 91470711},
-	1288743840: {103835610, 2061228961, 399631275},
-	1355427840: {2088165637, 92642117, 255365401},
-	1288746300: {693366042, 108949880, 511160300, 1112307290},
-	1325181360: {91150972, 91251780, 1807097639, 1114091017, 1112307290},
-	1328832060: {842914285, 1706597650, 1260530091, 235973308, 1679869203},
-	1333305840: {91834135, 1759260185, 90279117},
-	1340513220: {1388302843, 1308898048, 642805722},
-	1341872820: {91834135, 91930761, 90820598, 91988983},
-	1346789280: {91586558, 90128894, 90024369, 1384834734},
-	1354335840: {91838667, 91577863, 91577641, 230487602, 624121537, 90795389, 449406315},
-	1416163740: {93828578, 2023337958, 93809839, 93822543, 93089827, 90049792},
-	1457299560: {1789144460, 889021810, 91373426, 91392382, 206300515, 92235119},
-	1458437880: {91146293, 1728342768, 1404768079, 1038306400, 156051787},
-	1480387080: {90169405, 91559813, 91088102},
-	1483492380: {1860979659, 92366981, 92381231, 93187183, 93844721, 92068965},
-	1487537880: {94869322, 93858141, 95035410},
-	1289630940: {182783846, 1409616344, 982925111, 1663705146, 2071686927},
-	1327270260: {90545759, 90383111, 1383850848},
-	1300247820: {1499016488, 90307018},
-	1419339900: {91942581, 1350350854, 95079083, 90218862, 1098101066},
-	1320658440: {91270766, 327215256, 91251687},
-	1328667300: {91299034, 90920507, 572488132},
-	1341282840: {92187544, 92184456, 92184462},
-	1288795740: {900647791, 1140770188, 1433526867, 1971624443, 1352476124, 1388699646, 264805003},
-	1288779420: {598705413, 2053159977, 414416787},
-	1288914540: {1265833632, 985088173, 1550453404, 436571581},
-	1288993200: {1674676662, 2071073391, 1540419160},
-	1288761480: {626486972, 1121756094, 1537058401, 1059568892, 1641266158, 1194626835},
-	1288763340: {1308834578, 1141984025, 824326683, 1121680699},
-	1306064220: {1805909553, 90424488, 90603945},
-	1311421380: {90424488, 90603945, 90551343, 90495655},
-	1306048380: {1270489052, 1706487568, 1323111135, 1624024794},
-	1288873620: {1905959242, 1464179334, 994214022},
-	1308808080: {1182236347, 904184913, 90908630, 463034046},
-	1310339940: {168672455, 1637937515, 2014664896, 90810234},
-	1313366760: {2014664896, 1637937515, 90810234, 583331358},
-	1288781700: {671963485, 585867827, 282243238},
-	1289109360: {528253166, 441145841, 1980619514},
-	1351424880: {493241505, 1568393323, 1980768550},
-	1288789920: {555899210, 417998042, 1420876159},
-	1313934180: {1088403918, 1609167649, 1043751656},
-	1288794720: {1404371206, 1870415382, 311962535},
-	1288798080: {626196688, 1762147410, 914401657},
-	1340372220: {90791317, 90402600, 91590931},
-	1289115240: {735504329, 1824434236, 433380142},
-	1300429980: {1556318299, 278147861},
-	1354436520: {92297219, 90676292, 1986823588},
-	1364411100: {667730805, 92363727, 92799758},
-	1391082360: {90278191, 1246847589, 1835908064},
-	1288808820: {331306110, 334757740, 187294207},
-	1318162260: {331306110, 187294207, 1914276689, 334757740, 1340729910},
-	1377987060: {1699302872, 92905488, 93149064},
-	1293579540: {424221925, 496233015, 1615174181},
-	1289241540: {1270335593, 979712389, 1411283429, 848051766, 158238629},
-	1295666520: {123808541, 1270335593, 816572770, 979712389, 90000192},
-	1295720940: {175792799, 1087737999, 282317545},
-	1297128840: {816572770, 123808541, 220473985, 1270335593},
-	1309957020: {90384623, 90542730, 90509976, 90469420},
-	1310646420: {90424897, 90951929, 90613468, 90644177},
-	1316555400: {90068680, 90130165, 90914764},
-	1322421900: {1337305208, 1188262821, 1030065101},
-	1333394160: {1077979930, 1906510180, 808347912},
-	1350056820: {235424962, 664609135, 124536199},
-	1354114680: {979712389, 91297325, 90000192},
-	1354904520: {128377556, 90998607, 91072528},
-	1414016700: {90384330, 92889281, 91607146},
-	1288829940: {824960655, 1084235893, 297414143},
-	1288854180: {90005305, 1500583214, 1015412623},
-	1289045520: {607742738, 483337313, 1144606277},
-	1289988360: {1980078528, 725310764, 389739502},
-	1288832700: {779117283, 454355818, 1089333128},
-	1331018280: {1206213630, 570359170, 90956313},
-	1288865340: {423594466, 1938769991, 1733042569, 1839947683},
-	1288927140: {585138321, 833549019, 1581024045, 1323704625, 1931927955, 501836760, 1704564521, 358353304, 1524850555, 1327176493},
-	1288840440: {139203675, 861514240, 745725237, 1485709285},
-	1289198880: {278036129, 1563947604, 897565647, 1440256641},
-	1289374920: {708949913, 1825968244, 606789702},
-	1289459940: {1935264522, 984986181, 1533888437, 881371325, 1369152742},
-	1288932120: {1480925632, 1340873988, 1402544048, 1773603089, 1629247567, 535013425, 909833874, 673851748, 147784030, 150468894, 732537055, 1689418471},
-	1289331000: {303656160, 626957634, 1140016443, 1329202439, 301522928, 1815347836, 1967627916, 803592164, 886675379, 1143140371, 1791490447, 1095893585, 841618538, 1637290992, 171198191, 181965170, 262713025, 1554720782, 841673602, 487312118, 957671442, 1600529454, 616810247},
-	1289423880: {392136228, 245301667, 710506298, 194990795, 1022084952, 1614135877, 494351419},
-	1289500200: {430192870, 647939902, 975278934},
-	1289595660: {700441423, 1449466864, 161685342, 740455641},
-	1289660100: {1863399611, 1603470132, 1295209848, 853410349, 263303276, 272664229},
-	1289924400: {1539365358, 1619165511, 1244194432},
-	1290026940: {408373313, 1761724371, 1978006738, 852474487},
-	1290547200: {1045497610, 2062814796, 684576134, 1672036418},
-	1290682020: {758324030, 1067110795, 2046965124},
-	1291487400: {736681405, 147702594, 175279874, 1137146327},
-	1293917580: {716199683, 223252712, 1452715302, 1608251190},
-	1295227680: {1095893585, 1791490447, 1931124236},
-	1295262420: {1047416303, 591212593, 714579949},
-	1303157400: {664924026, 125637685, 1191363856, 1750394674, 1093637842, 1727764995, 626411552, 90061065, 205216311, 901320515, 771499021},
-	1303222140: {304127292, 1356298321, 1159571872, 1255320531, 464110869, 876632496},
-	1303318200: {1345849360, 1039026218, 150468894, 1689418471, 1303872973, 527664984},
-	1303423800: {144889577, 148464538, 1508894881, 1698941081, 192941486},
-	1303518900: {1610268506, 1913819309, 844792244},
-	1303745700: {1406255952, 1009694287, 412503733, 649224051, 853583291, 910482587},
-	1303838400: {123561032, 1472370020, 614051529, 1349690510, 636272125, 144208395},
-	1303922700: {830541135, 247451439, 890939563, 106917400, 1336967147, 1394733798, 219741648},
-	1303993920: {1397278285, 1608251190, 909833874, 844646062, 1722195070, 261618456, 227508484},
-	1304163420: {2061075482, 315083056, 1765381384},
-	1304333460: {1277962552, 1325314214, 90551936},
-	1304442720: {1567632405, 1508260954, 1320247443, 864199971},
-	1307122140: {152829791, 781331374, 1967733046},
-	1309739940: {961076244, 835257191, 1291106522},
-	1310144760: {908144793, 223727748, 365675114},
-	1312351020: {1208288405, 1440435214, 499363470, 436502473, 1907517637, 1509907757, 1218854180, 1716720509},
-	1312667880: {254402328, 984737980, 1638000636},
-	1312801860: {695511828, 1827350398, 1095893585, 1791490447, 421392520, 1931124236},
-	1313625900: {957671442, 1554720782, 487312118, 1573027162, 841673602},
-	1316039820: {594067954, 630343471, 403759464},
-	1288881300: {963393540, 1463122254, 413191808, 1321565044, 407890861, 1247507500, 894696471},
-	1288903560: {1321132601, 1321755847, 220353378, 1238956593, 987838065},
-	1309758360: {90958887, 90958838, 90958933},
-	1288847220: {578671306, 787257304, 1284937280, 797420437},
-	1288847340: {748382603, 1046800359, 1952525621},
-	1288847580: {1261313867, 991935578, 220081727, 1306254180, 849106728},
-	1288848660: {1663088330, 1117707046, 1572041388, 659086385, 1594102773, 496287066},
-	1292373960: {542781313, 90002709, 496287066},
-	1308979860: {1349566949, 496287066, 975930206, 90612411, 1197612362},
-	1382739120: {90444791, 93322386, 91114952},
-	1288850520: {1053105037, 322389791, 145199643, 1333748060, 1930429402, 182777259, 296125285, 1345997634, 539096076, 561607365, 190400728},
-	1289234400: {1512187101, 143701059, 872998027, 486722920, 762723689, 660205036, 1586041140, 554951164},
-	1289314380: {1226392581, 1463255484, 750258878, 1796071892, 2041507547, 606566302, 1984259785, 685450490, 90025057, 90025035, 415265456, 1639861607},
-	1289368980: {643820926, 1504191548, 1773945579, 2019321103, 294078445},
-	1309081920: {1733650021, 90870508, 90872593, 90903119, 90912145},
-	1289959980: {1872660286, 1770215525, 1495266333},
-	1291711320: {314168553, 1872660286, 1331728111},
-	1291341660: {90027306, 745601066, 1477797473, 90104041},
-	1292829480: {90167492, 90166366, 90170955, 90172723},
-	1288860720: {1140616637, 1142841650, 849486836},
-	1288987080: {1382936392, 1636081714, 1544474176},
-	1288862280: {1425293347, 804920019, 1613493125, 131650259},
-	1307456040: {780968076, 1149089908, 1410444601, 1289560308, 406162202},
-	1307514660: {364146051, 1183603907, 715152096, 90565926, 1613493125},
-	1339315320: {92121038, 92071921, 92112044},
-	1339859340: {273237918, 91981065, 92055317},
-	1340097060: {92138017, 92115241, 92144348},
-	1340350140: {92148484, 92158229, 92110541},
-	1341986580: {92234446, 92228834, 92231799},
-	1342462860: {92223820, 92231316, 92223966},
-	1342511160: {92245647, 92009911, 92235241, 92256630},
-	1344570180: {763745671, 91955507, 92334175},
-	1346284740: {92379518, 732147348, 92388957},
-	1349640780: {92501692, 92450878, 92475231},
-	1349837160: {92516241, 92501847, 92496872, 92404792},
-	1350121440: {92490030, 92462765, 92463771, 92513893},
-	1353263820: {715152096, 92256438, 91184877},
-	1355704560: {92672600, 91763994, 91499947},
-	1355828280: {92538581, 92489634, 92628832},
-	1356142740: {784290460, 92727536, 778625673},
-	1357790820: {1131933691, 91692001, 92754837},
-	1358410800: {92804372, 92546333, 92717862, 92615958},
-	1289731020: {356330836, 1164368445, 835994957, 1198930535},
-	1294186140: {835994957, 1198930535, 90063184, 809445394},
-	1306299900: {90381163, 90521222, 90498342},
-	1307945460: {90498342, 90521222, 90381163},
-	1288866180: {1123130654, 2084957016, 1103256560},
-	1290933420: {496640160, 1831786388, 1339879351, 1863113745, 2051457455},
-	1351691040: {197971224, 90007440, 90008070},
-	1288892520: {1075453929, 1459724934, 922765974, 1480533127},
-	1289157840: {340786109, 363063164, 1753923549},
-	1289897160: {1510351491, 1168357321, 1778980377},
-	1288895340: {704795161, 1809547022, 1295315729},
-	1288976880: {909594410, 852141971, 483831334, 1924545336},
-	1292370480: {1559226705, 345522171, 117074457},
-	1292529120: {1502639267, 1470903133, 660196112},
-	1292677140: {1969368319, 1756742247, 1388610778},
-	1294686000: {570385439, 658151612, 730618358},
-	1294928040: {358634130, 1142035326, 1490128261},
-	1296092640: {1770044067, 762380105},
-	1298505240: {964175720, 1421066538, 1585193979},
-	1298565180: {619034417, 90444053, 870743612},
-	1298989500: {1735374354, 787034971, 1135600728, 960602310},
-	1299078960: {1106841393, 100792143, 90453956, 90406084, 90396598, 90356160, 960602310},
-	1299170100: {1297379475, 1948527131, 90160392, 90473461},
-	1299257520: {704346737, 1549975406, 1792924439},
-	1299375600: {697794219, 626392799},
-	1299598500: {144902393, 90461152, 90282639, 90331454, 90204920, 90177341},
-	1299942420: {90453956, 2043216776, 1291016411},
-	1300236840: {90356227, 1336564805, 1544068060},
-	1300442640: {996937897, 635164602, 314513554, 993978156, 960952754},
-	1300584420: {90498347, 618796242, 1627965855},
-	1300892700: {1945100765, 90041041, 1516571004, 1857186134},
-	1301354640: {1265833632, 90299937, 1151555985},
-	1301518620: {1002863089, 1858222156, 1228058745, 1145464524, 90150892, 2005819283},
-	1301654880: {378467968, 1973526666, 904810258, 1706426585},
-	1301754840: {1418863988, 1453455117, 350425235},
-	1301838540: {460555112, 1561011491, 858183691, 1037880612},
-	1301923320: {413485920, 613370476, 90272153, 685566788, 283254187, 847733584, 90218902, 1021577884, 90497924},
-	1302010440: {980235502, 756807665, 1253496127, 90364990, 1637675461},
-	1302088680: {247108131, 384124719, 1636816335, 1811124177, 1701992388},
-	1302188580: {90612658, 90062908, 490878306},
-	1302362760: {90212956, 1553887153, 90319586, 880945694, 1539836283, 2011681081, 900946287, 1397921604},
-	1302428700: {1788224590, 1462400718, 1450483557, 802285139},
-	1302533760: {161211018, 1268080790, 90449903},
-	1302641100: {90560509, 90051075, 1514553708},
-	1302697980: {90445298, 408545758, 2061181883},
-	1302873660: {90478152, 90534830, 1038844498, 1509733736, 888694330},
-	1302950520: {911450294, 90449903, 1412763801, 1408159802, 755995397},
-	1303222980: {90063992, 986412032, 736865087, 1040288570, 180237465},
-	1303381920: {172627229, 2011681081, 1397921604},
-	1303642560: {2003063259, 1349566949, 1234657739},
-	1303990800: {90482018, 90689420, 90547008},
-	1304518260: {90449903, 265046126, 1757984406},
-	1304842320: {1661367243, 1420452470, 465987924, 358591855, 912543213, 1508520407, 183834119},
-	1305101880: {1738966258, 164865933, 1852658638},
-	1305225120: {90649630, 1355866824, 861829964, 1247503569},
-	1305725100: {436582662, 1598649612, 1514498198, 447333425},
-	1308964920: {198788738, 1485095741, 1692750878},
-	1309660440: {900946287, 1539836283, 880945694},
-	1310166420: {1393924252, 1574582885, 1664876572, 761783438},
-	1311652560: {192605020, 1147798128, 402623482},
-	1312282500: {1924444439, 911967810, 1542338653},
-	1313140140: {772715452, 90905306, 1400724261, 90912851, 922257905},
-	1313436240: {1007393046, 1437091803, 1475444478},
-	1314710280: {1313120774, 124738076, 90964636},
-	1314811920: {1431724153, 1101072164, 443074620},
-	1315347540: {696067142, 1782852060, 1540237728},
-	1324547520: {90671371, 362098950, 618497069},
-	1326805560: {90375718, 899737931, 90895799},
-	1328515260: {911967810, 90056398, 90538520, 1694713273, 125103937},
-	1329155880: {589373202, 692134104, 1957433932},
-	1333280340: {90895883, 181541104, 91309751, 91580724},
-	1339918620: {91018218, 90211047, 90194561},
-	1340002620: {626392799, 1809547022, 90376350, 911450294, 377290319},
-	1340172960: {265046126, 1575672892, 1525565859},
-	1342329900: {91309751, 90649630, 618497069},
-	1346149200: {1734129574, 1696603755, 90895799},
-	1347621060: {1040288570, 932714118, 1349566949},
-	1348290300: {2024841657, 90615058, 1734129574, 90253808, 91495422, 1342165595},
-	1354747620: {92010272, 1886666289, 231773171},
-	1354873020: {1646608441, 719599149, 1676578919},
-	1355146980: {244127454, 1868415009, 1028712811, 1326740956, 1166724998, 1455975253, 91500914, 91500951},
-	1355351100: {92300422, 91445008, 1997378609, 369642251},
-	1359199680: {90498347, 719599149, 1676578919, 1456314982, 1646608441},
-	1371967860: {93362142, 683897610, 1456314982},
-	1372708260: {2002110761, 840702920, 790783979},
-	1373899800: {1924444439, 1377875264, 1302876313},
-	1378296180: {1105656888, 1889736658, 93344879},
-	1378365480: {716630529, 92558085, 92399393, 92541471, 90425634},
-	1378618740: {808160074, 90056398, 1873504478, 144715471},
-	1378705080: {675075027, 1924444439, 1809011971, 857437635},
-	1379035260: {1909892190, 92216465, 91988762, 879584101},
-	1379884860: {90336981, 90041993, 90468520},
-	1379969160: {1582372841, 616200675, 636078813},
-	1380654300: {90508357, 1021030278, 91592687},
-	1381958040: {1809547022, 91753554, 90315602},
-	1383804660: {807300079, 1242011384, 800079953, 90036110, 858975070},
-	1383980340: {1579185973, 827115187, 1629442181},
-	1384063380: {90999823, 93799263, 93833237, 90187014, 1788224590, 91113574, 93833317, 93831472},
-	1384508340: {91248032, 90423056, 1167475273, 93201672},
-	1384587180: {91247989, 91442777, 91438971, 299862503, 1403154436},
-	1385915100: {1809547022, 91753554, 1735374354},
-	1386458100: {93974288, 91263305, 1699611261, 90259063},
-	1387996740: {93833317, 90579159, 90768985},
-	1393769280: {91333750, 1421680088, 422209603},
-	1394213280: {277233044, 424740857, 184393782, 736565888, 1408383372, 176182070},
-	1394345280: {91384710, 90294263, 91753778, 370334050, 92614310},
-	1394562600: {295867462, 91242804, 353073162, 90934126, 1159810783, 90101265, 877669878, 93169817},
-	1394618640: {496104507, 1892267588, 92956906, 1244913262, 91357110},
-	1394688120: {837845819, 93376776, 91456292, 93034624, 688581994},
-	1395034440: {618230958, 1272201704, 838412410, 872372951, 552771202},
-	1395174960: {204767116, 1142571084, 93257234},
-	1395467940: {838412410, 93110866, 91976829},
-	1395638580: {1375317674, 643434767, 911967810, 91753554, 1809547022, 1193073790, 220566298},
-	1395766320: {650662730, 636057070, 1535383889, 92036370},
-	1395986400: {605838897, 91105581, 92360447},
-	1396290240: {94072449, 92959039, 90711670},
-	1396364940: {909869502, 91472880, 92634668, 91838279},
-	1396696260: {92093788, 1319712638, 92511267},
-	1396777020: {296512013, 1695248452, 1184413832, 92685201, 309390849, 93996401, 617207885, 653411463},
-	1397129160: {187987599, 1240188226, 92184773},
-	1397199240: {815792248, 1966453282, 90577194, 557877097, 90291384, 92733393},
-	1397281320: {93279215, 419575650, 93902855, 1480228907},
-	1397416320: {128978357, 1049773079, 93508696, 662202161},
-	1397546100: {1472725775, 92862781, 789633930},
-	1397801700: {492982062, 94507466, 91801482, 91385920, 740263119, 94420328, 1574766938},
-	1397887980: {1733222144, 1500500714, 94182427, 1500165367, 390651764, 1289076812, 1843638539, 91082697, 94194089, 94173529, 1676766295, 91779246, 564175413, 1662935446, 94181728, 1290100870, 94276182, 91059719},
-	1397970840: {93823636, 911778693, 93871674, 635654594, 92030692, 93813546, 93788263, 93964846, 94165911, 92544466, 560874264},
-	1398757800: {91672357, 91813980, 91678523, 92497122},
-	1398872580: {1852145323, 91442575, 1148361511},
-	1399093320: {2011991119, 1456314982, 1646608441, 92007552},
-	1399211160: {261421576, 90552503, 90650835, 94396095, 94396128, 94396105, 91761078},
-	1399851000: {92197269, 219371166, 1683778741},
-	1399964580: {719599149, 90934126, 94131852},
-	1400332920: {91936817, 90711252, 91808567, 92800615},
-	1400394120: {154002750, 1395725360, 377405587, 91511569, 91281221},
-	1400548920: {1864901734, 90622608, 90795993},
-	1402365420: {90753573, 92180211, 92774815},
-	1402817040: {1374728923, 94121410, 1167475273, 91239910, 1146383403, 659979371},
-	1402989660: {2012019678, 94072449, 982973787},
-	1403848500: {91505772, 175561979, 91011416},
-	1404628500: {93999865, 92250557, 94649378},
-	1405111140: {1198763894, 119375229, 469638354, 277430685, 1442661262},
-	1406124180: {93770194, 93770508, 93165840, 92325381, 94435389},
-	1407086700: {874274208, 765600781, 90067467},
-	1441882560: {1629442181, 90711252, 1141738106, 93376776},
-	1455289320: {1629442181, 93860671, 1141738106},
-	1295083680: {1178538365, 1441959667, 657464545},
-	1339760580: {1178538365, 657464545, 1441959667},
-	1353307380: {657464545, 1178538365, 1441959667},
-	1292282760: {1883583160, 936974757, 90009318},
-	1295903760: {742260700, 839199979, 90329898, 90337819},
-	1291171920: {1037937708, 177984669, 909428716, 1082169200, 530803567, 1473423443},
-	1291339440: {1778093453, 115576569, 1251223357, 1473341886, 1831033716, 1732723631},
-	1297103880: {90352328, 1309063150, 1849640281, 124886999},
-	1288904340: {154090272, 338078526, 90002806},
-	1288992480: {90006965, 584644002, 1283833429, 581928331},
-	1309045980: {2035158907, 1827504816, 90771779, 90878213},
-	1309116240: {1656634005, 90755762, 771802962},
-	1331019060: {90502098, 1017630217, 224086051},
-	1289000460: {859583385, 1865942684, 873602051},
-	1289072700: {730564400, 1128977205, 582439838, 1588872713},
-	1288911540: {1197250097, 353543193, 1644513477},
-	1288993320: {1586533074, 1288566949, 1218083158},
-	1289050560: {570658548, 178584643, 2077414435},
-	1294405980: {1429973844, 90214696, 1436901509},
-	1297898040: {90031224, 90210954, 90338585},
-	1356035700: {91967925, 673192319, 278308418},
-	1373720880: {366891342, 661012730, 1414804516},
-	1295431020: {90178583, 1117440769, 90229869},
-	1307178540: {90783725, 90783776, 90783649},
-	1289066580: {952987707, 964387670, 1138872170},
-	1289038440: {90011079, 90012024, 90014091},
-	1294430340: {90040614, 90098240, 90117145},
-	1304067360: {1984253923, 90341114, 90683634},
-	1304165280: {90686221, 90686156, 90696619},
-	1304406000: {1516150141, 90705835, 90679025},
-	1305247200: {90720527, 90736240, 1621436246, 1326668737},
-	1305270180: {1621436246, 1326668737, 90736240},
-	1292869200: {1818196718, 410166959, 1158907493, 1556153804, 846552781, 1293983527, 469761444, 363969650},
-	1292915580: {846494162, 1048122391, 526442597, 1428350086, 1181017136, 936023232},
-	1292998980: {949192361, 879618999, 733976026, 1003273528, 1149676059},
-	1296067080: {1908654067, 1794863977, 815334570},
-	1298885520: {946867729, 1429716033, 90079017, 1129092353},
-	1300058640: {551972491, 435846426, 1848729766, 737120536, 396610112, 1427716302},
-	1300152180: {753000622, 1664534935, 970991624, 933961532},
-	1300184880: {2035047876, 1665533067, 764402719, 408663162, 358372318, 953088554, 153864464, 318662182},
-	1302260040: {611854715, 815334570, 318662182},
-	1302910500: {1317976564, 1396370446, 1052226829},
-	1303031340: {1901572700, 433302027, 529215950},
-	1305263160: {292483157, 312290624, 759104218, 1801567197},
-	1305959760: {90006771, 159501023, 1804388257, 924121230},
-	1306743660: {90805835, 1857542410},
-	1310721420: {2056464820, 1046512625, 1148620708},
-	1315476180: {526442597, 136901084, 1148620708},
-	1377324540: {849643996, 351003049, 739245004},
-	1402876440: {90008118, 944142273, 1498605733},
-	1288934400: {658701538, 1834813492, 1831637303},
-	1332888600: {899903948, 638132473, 1621035860},
-	1380767940: {1615253403, 92988105, 93318114, 93425882},
-	1380862920: {90329942, 90816680, 91540894},
-	1381185900: {1410698854, 90613846, 93432371},
-	1289124600: {435252660, 2006684307, 1234172163},
-	1390039680: {93654551, 978025598, 290813713, 93659105},
-	1289210760: {567583729, 769925221, 708338672, 1205714525, 923812488},
-	1289911140: {153712083, 1240443687, 2002722359},
-	1292055240: {1627351341, 90090139, 1190341940},
-	1294308960: {1631190985, 155105840, 2030690181},
-	1288959480: {1019789950, 90001814, 923482227, 108638615, 90011232, 146299674},
-	1289226480: {683615416, 1067570807, 90022595},
-	1289330760: {90024186, 521625785, 905084720},
-	1289486520: {282925672, 692158018, 1724722397},
-	1288975620: {2048841430, 925930367, 1978622450},
-	1291800240: {90119030, 553196637, 1713378278, 1578832434, 90062112, 299134653},
-	1291882800: {820691860, 90082040, 1295168457, 1433832403},
-	1289311440: {409252918, 1721864142, 388099088, 678880762, 972123695, 1132564973, 341586003, 520908763, 1160815716, 1105152931},
-	1289421660: {90027166, 666977633, 1984392727, 159367143},
-	1289933880: {651004893, 270915210, 552549383, 818929081},
-	1291578600: {727745290, 131102853, 1689380834},
-	1294695900: {1507939610, 1721864142, 1716245277, 1764117673, 1266139951, 384768590, 145587484, 341586003, 520908763, 746447320, 746888873, 131102853, 90099676, 1132564973, 1582562101, 495990105, 350145435, 1194383922, 1813906902, 1615849669, 90213249},
-	1294728960: {90027166, 1812116498, 1353044171, 1042541029, 1131967570, 338964832, 758966296, 1200690114, 90173746, 904958126, 90234516, 263012491, 1256566428, 409252918, 110653176, 517047208, 1105152931},
-	1294847460: {90158922, 90147850, 2047472648, 1185794676, 979308798, 116480354, 305233823, 1880529709, 90121265, 1958101768, 90217749, 1350059433, 1722605646, 90249433, 1351906514, 1138532516, 307191430},
-	1294919100: {972123695, 1347352343, 1198367481, 989276596, 653256513, 355928159, 1178421148, 1351906514, 818929081},
-	1295030340: {1343136765, 90218723, 390223785},
-	1295071560: {1689380834, 417262639, 90167056},
-	1295721000: {1198367481, 1716245277, 989276596, 390223785},
-	1296519840: {979308798, 90194544, 116480354, 90086327},
-	1296703920: {972123695, 110653176, 1347352343, 1000542856},
-	1298118720: {90184126, 263012491, 1374506715},
-	1299089160: {814921855, 746447320, 307191430},
-	1303355160: {1336686499, 424363419, 90343612},
-	1306171920: {90395857, 90273645, 90086327},
-	1307384940: {746888873, 90099676, 424363419, 966783736, 90634036, 90710578, 90596190, 90579434},
-	1307452080: {307191430, 90616429, 90661608, 116480354},
-	1307551440: {730269637, 904958126, 90299417, 1561202771, 90395857, 1544396049, 1299495154, 961269879, 1208191865},
-	1307649780: {2047472648, 1131967570, 1042541029, 240462255},
-	1308149760: {550633829, 90706789, 1401474335, 130191784},
-	1309371840: {287080206, 302296038, 483951246, 90059013},
-	1310041080: {1165459679, 957839657, 90596190},
-	1311337800: {1351906514, 145587484, 90086327},
-	1311423600: {550633829, 1336686499, 90596190, 2047472648, 90288779},
-	1311597660: {90184126, 1812116498, 2088345565, 730269637, 1105152931},
-	1311687600: {1435879173, 1911639727, 1134117687, 1531439483},
-	1313081100: {2007491999, 1034670053, 1709722973},
-	1313618820: {989276596, 145587484, 1911639727, 263302224},
-	1315937100: {1814238541, 234968379, 90977061},
-	1317995940: {904958126, 90299417, 838425425, 91218108, 1967001659},
-	1318179540: {2047472648, 90299417, 838425425},
-	1320185820: {424363419, 156186939, 1194383922},
-	1320750720: {1077761965, 813290359, 814921855},
-	1328719920: {90828665, 1742153816, 91178054, 91081166},
-	1329927060: {1077761965, 783590679, 839117974},
-	1334846880: {1350059433, 1927064512, 941797016},
-	1337059380: {817581135, 117146050, 706287316, 1014039255, 91212357},
-	1344521880: {1065202196, 350145435, 643259853},
-	1346729880: {2088345565, 1401474335, 730269637, 90724229},
-	1347031200: {1142724364, 91675442, 91605581, 1177485512, 90868863, 90738711},
-	1347203760: {90925320, 91524338, 90738841, 90708034, 1813906902},
-	1351473000: {92277125, 1164552158, 92402787, 91172874, 90813771, 92011662, 726216902, 984900780, 92536396, 1343377510, 92233538, 90932296, 558580565, 91781031},
-	1352070240: {92384212, 92570758, 90909582, 91753442, 91979886, 91942860, 91776150, 91754830, 91769775, 91466342, 90192005, 92585250},
-	1352310360: {92573189, 92568270, 1507939610, 920984912, 972123695, 230426272, 1628288954},
-	1352522340: {91589019, 426301685, 995059348, 91484660},
-	1353191340: {904958126, 91218108, 90299417, 838425425, 91197512},
-	1354188180: {818929081, 770015296, 240488997},
-	1355190480: {92422870, 757412022, 92431715},
-	1357271880: {908185315, 752941466, 676513851},
-	1360627560: {583177990, 92781102, 92761492, 92670785},
-	1360977180: {92437227, 92708567, 90121265, 92939113, 92851262, 92149658, 1142501010},
-	1362839700: {272520470, 91468237, 90901180},
-	1363368120: {966783736, 1067521336, 91470167, 359282159, 90182076, 92952623},
-	1365960420: {92830175, 92518351, 1011830348, 90229123, 1958101768, 92851262},
-	1366371180: {364048534, 746447320, 91579375},
-	1367156340: {91381615, 920984912, 90921829},
-	1367319420: {93249066, 1821988878, 91439056},
-	1380797460: {92591976, 90271205, 90147850},
-	1288984020: {1233925437, 1849570700, 1810028952},
-	1289070360: {1930290603, 1603647786, 941408197, 858393299},
-	1289170740: {608753819, 1998120351, 850263889, 882309291},
-	1289611200: {149871941, 1243800371, 1449536932},
-	1290878880: {1814872989, 1549427172, 1814293548},
-	1289392500: {514985820, 1008871180, 704704451},
-	1289498520: {1347567520, 808100592, 2007483033},
-	1291936620: {967026511, 314914921, 90125399},
-	1292976840: {90181672, 1541411069, 1811578151},
-	1295570700: {1127611532, 1058675140, 1626148671},
-	1300035720: {90346329, 975785791, 90132574},
-	1308871980: {90063981, 90878203, 1077883165},
-	1315263060: {90898573, 1459671461, 1215251249},
-	1316900880: {870696354, 740006500, 880192859, 90882404},
-	1319125020: {712490350, 90654795, 1485320437},
-	1339826340: {91491905, 91562955, 971376567, 164876242},
-	1340960040: {446855701, 91985806, 92014263},
-	1289168820: {1053771841, 1763523411, 120280120},
-	1289066760: {199621007, 917479950, 1809690643, 133888864, 850263889},
-	1289157420: {1421066538, 1717149436, 901026579},
-	1298571060: {608753819, 677482381, 1809690643, 1681050140, 223928130, 1421066538, 1416896499},
-	1289119800: {516959764, 1694763431, 1959072424},
-	1289703060: {800228943, 1116284449, 334793642},
-	1289163360: {736158920, 712534537, 1006919828, 154913831},
-	1289238420: {307980762, 384351439, 1827790327, 90000184, 1170137165},
-	1289002020: {216892474, 414075386, 540717793},
-	1289331840: {90010134, 1701826206, 1598878208},
-	1289625300: {1461277341, 334234504, 582489390},
-	1289628720: {1976459018, 1494533517, 410495060, 174761544, 1057150064, 442206309},
-	1291874580: {1753517390, 562301466, 90131287},
-	1289116260: {536077683, 1836116133, 2069678145},
-	1289203260: {396131806, 979392583, 1631005420},
-	1289148660: {1181063468, 2034693763, 1771729830, 2081588157, 1611226195},
-	1294179600: {90146352, 90021845},
-	1291517460: {90117507, 90118504, 90090570},
-	1300690200: {114074271, 432597349, 551908159, 1301823499},
-	1300856580: {1544068060, 1199790290, 90218032, 90212420},
-	1301736420: {341587888, 147373558, 853088748, 526282137, 1780001370},
-}
