+++ internal/redisqueue/queue_test.go
@@ -1,0 +1,51 @@
+package redisqueue
+
+import (
+	"testing"
+
+	"github.com/antihax/evedata/internal/redigohelper"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestHQ(t *testing.T) {
+	pool := redigohelper.ConnectRedisPool([]string{"127.0.0.1:6379"}, "", "", false)
+	hq := NewRedisQueue(pool, "test-redisqueue")
+	err := hq.QueueWork(
+		[]Work{
+			Work{"alliance",
+				2,
+			},
+			Work{"alliance",
+				3,
+			},
+			Work{"alliance",
+				4,
+			},
+			Work{"alliance",
+				5,
+			},
+			Work{"alliance",
+				6,
+			},
+		},
+	)
+	assert.Nil(t, err)
+	size, err := hq.Size()
+	assert.Nil(t, err)
+	assert.Equal(t, 5, size)
+
+	var work []*Work
+
+	for i := 0; i < 5; i++ {
+		w, err := hq.GetWork()
+		assert.Nil(t, err)
+		work = append(work, w)
+	}
+
+	check := map[int]bool{2: true, 3: true, 4: true, 5: true, 6: true}
+
+	for i := range work {
+		delete(check, work[i].Parameter.(int))
+	}
+	assert.Empty(t, check)
+}
